refactor PSCi code
