+++ psci/PSCi/Directive.hs
@@ -15,6 +15,10 @@
 
 module PSCi.Directive where
 
+import Prelude ()
+import Prelude.Compat
+
+
 import Data.Maybe (fromJust, listToMaybe)
 import Data.List (isPrefixOf)
 import Data.Tuple (swap)
